# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HEXA Purchase Assistant API
  version: 0.1.0
  description: |
    FastAPI backend for Flutter app, WhatsApp webhook, and Super Admin.
    AI is used for parsing only; business logic stays server-side.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /v1/businesses/{businessId}/entries:
    get:
      tags: [entries]
      summary: List entries
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: item
          in: query
          schema:
            type: string
        - name: supplier_id
          in: query
          schema:
            type: string
            format: uuid
            description: Filter entries by supplier
      responses:
        '200':
          description: Paginated entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryListResponse'
    post:
      tags: [entries]
      summary: Create entry (preview or confirm based on body)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
  /v1/businesses/{businessId}/entries/{entryId}:
    get:
      tags: [entries]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
    patch:
      tags: [entries]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryPatch'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
  /v1/businesses/{businessId}/entries/parse:
    post:
      tags: [entries]
      summary: AI parse draft only — returns structured fields for user confirmation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseDraftResponse'
  /v1/businesses/{businessId}/entries/check-duplicate:
    post:
      tags: [entries]
      summary: Check duplicate item+qty+date
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DuplicateCheckRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateCheckResponse'
...
  /v1/businesses/{businessId}/price-intelligence:
    get:
      tags: [price-intelligence]
      summary: PIP for item + optional current price
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item
          in: query
          required: true
          schema:
            type: string
        - name: current_price
          in: query
          schema:
            type: number
        - name: window_days
          in: query
          schema:
            type: integer
            default: 90
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceIntelligence'
...
  /v1/businesses/{businessId}/media/voice:
    post:
      tags: [media]
      summary: Voice/STT preview — no auto-save
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transcript preview
  /v1/businesses/{businessId}/media/ocr:
    post:
      tags: [media]
      summary: OCR preview — no auto-save
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OCR text preview
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntryCreateRequest:
      type: object
      required: [entry_date, lines]
      properties:
        entry_date:
          type: string
          format: date
        supplier_id:
          type: string
          format: uuid
        broker_id:
          type: string
          format: uuid
        invoice_no:
          type: string
        transport_cost:
          type: number
        commission_amount:
          type: number
          minimum: 0
        confirm:
          type: boolean
          description: If false, run validation only (preview)
        lines:
          type: array
          items:
            $ref: '#/components/schemas/EntryLineInput'
    EntryLineInput:
      type: object
      required: [item_name, qty, unit, buy_price, landing_cost]
      properties:
        item_name:
          type: string
        category:
          type: string
        qty:
          type: number
        unit:
          type: string
          enum: [kg, box, piece]
        buy_price:
          type: number
        landing_cost:
          type: number
          description: Manual landing cost (per PRD)
        selling_price:
          type: number
    Entry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        entry_date:
          type: string
          format: date
        supplier_id:
          type: string
          format: uuid
          nullable: true
        broker_id:
          type: string
          format: uuid
          nullable: true
        invoice_no:
          type: string
          nullable: true
        transport_cost:
          type: number
          nullable: true
        commission_amount:
          type: number
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/EntryLine'
    EntryLine:
      allOf:
        - $ref: '#/components/schemas/EntryLineInput'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            profit:
              type: number
    EntryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
        next_cursor:
          type: string
    EntryPatch:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/EntryLineInput'
    ParseDraftResponse:
      type: object
      properties:
        draft:
          $ref: '#/components/schemas/EntryCreateRequest'
        missing_fields:
          type: array
          items:
            type: string
        confidence:
          type: number
    DuplicateCheckRequest:
      type: object
      required: [item_name, qty, entry_date]
      properties:
        item_name:
          type: string
        qty:
          type: number
        entry_date:
          type: string
          format: date
    DuplicateCheckResponse:
      type: object
      properties:
        duplicate:
          type: boolean
        matching_entry_ids:
          type: array
          items:
            type: string
            format: uuid
...
    PriceIntelligence:
      type: object
      properties:
        item:
          type: string
        avg:
          type: number
        high:
          type: number
        low:
          type: number
        trend:
          type: string
          enum: [up, down, flat]
        position_pct:
          type: number
        last_price:
          type: number
        frequency:
          type: integer
        confidence:
          type: number
        supplier_compare:
          type: array
          items:
            type: object
            properties:
              supplier_name:
                type: string
              avg_price:
                type: number
...